refactor(local-ci): extract io_utils.py for I/O + locking helpers
R2-1 fourth cut. Move 5 helpers (tail_lines, trim_line,
atomic_write_text, image_change_summary, file_lock) plus the
LockBusyError exception out of local_ci.py into a sibling
io_utils.py module.

- local_ci.py: 11,033 → 10,961 lines (-72 net).
- io_utils.py: 110 lines.
- LockBusyError moves with file_lock (its only raise site); the two
  external catch sites in local_ci.py keep working via the re-export.
- image_change_summary keeps the Pillow-optional fallback so the test
  suite remains runnable on stripped CI bot images.
- io_utils depends on state_paths.ensure_state_dirs() — the first
  cross-module dependency inside the local-ci package, validating
  the seam choice.

Tests: 213/213 local_ci tests pass.
